+++ cfsmextract/func.go
@@ -1,12 +1,12 @@
-package main
+package cfsmextract
 
 import (
 	"fmt"
 	"go/types"
 	"os"
 
-	"github.com/nickng/dingo-hunter/sesstype"
-	"github.com/nickng/dingo-hunter/utils"
+	"github.com/nickng/dingo-hunter/cfsmextract/sesstype"
+	"github.com/nickng/dingo-hunter/cfsmextract/utils"
 	"golang.org/x/tools/go/ssa"
 )
