+++ cfsmextract/sesstype/generator/cfsm.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/nickng/dingo-hunter/sesstype"
+	"github.com/nickng/dingo-hunter/cfsmextract/sesstype"
 )
 
 var (
