+++ cfsmextract/sesstype/op_string.go
