+++ cfsmextract/utils/defs.go
