+++ cfsmextract/utils/emptyvalue.go
